# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API for demonstration purposes.",
        "title": "My API",
        "contact": {},
        "version": "1.0"
    },
    "host": "8.218.102.112:4880",
    "basePath": "/api/v1",
    "paths": {
        "/api/v1/user/me": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get details of the currently authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "Get user details",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
        "/user/message": {
            "get": {
                "description": "Get a message that needs to be signed for wallet authentication",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Get message for wallet authentication",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Solana wallet address",
                        "name": "address",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/user/wallet-login": {
            "post": {
                "description": "Authenticate user with wallet signature",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "authentication"
                ],
                "summary": "Wallet login",
                "parameters": [
                    {
                        "description": "Wallet authentication input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.WalletAuthService"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/ws/kline/{tokenAddress}": {
            "get": {
                "description": "通过 WebSocket 实时推送 K 线数据",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "WebSocket"
                ],
                "summary": "K线数据 WebSocket 接口",
                "parameters": [
                    {
                        "type": "string",
                        "description": "代币地址",
                        "name": "tokenAddress",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "1S",
                            "1",
                            "5",
                            "15",
                            "60",
                            "240",
                            "720",
                            "1D"
                        ],
                        "type": "string",
                        "description": "K线周期 (1S/1/5/15/60/240/720/1D)",
                        "name": "resolution",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "101": {
                        "description": "切换到 WebSocket 协议",
                        "schema": {
                            "$ref": "#/definitions/response.KlineData"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "response.KlineData": {
            "type": "object",
            "properties": {
                "close": {
                    "type": "number"
                },
                "high": {
                    "type": "number"
                },
                "low": {
                    "type": "number"
                },
                "open": {
                    "type": "number"
                },
                "timestamp": {
                    "type": "integer"
                },
                "volume": {
                    "type": "number"
                }
            }
        },
...
        "response.Message": {
            "type": "object",
            "properties": {
                "chain_id": {
                    "type": "integer"
                },
                "domain": {
                    "type": "string"
                },
                "expiration_time": {
                    "type": "string"
                },
                "issued_at": {
                    "type": "string"
                },
                "nonce": {
                    "type": "string"
                },
                "statement": {
                    "type": "string"
                },
                "uri": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "response.MessageResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/response.Message"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "service.WalletAuthService": {
            "description": "用于钱包认证和登录的服务结构体",
            "type": "object",
            "required": [
                "address",
                "signature"
            ],
            "properties": {
                "address": {
                    "type": "string",
                    "example": "5YNmS1R9nNSCDzb5a7mMJ1dwK9uHeAAF4CmPEwKgVWr8"
                },
                "signature": {
                    "type": "string",
                    "example": "base58_encoded_signature"
                }
            }
        }
    }
}